fix: output dir drop /bin (#52)

+++ scripts/build_grpc.sh
@@ -3,6 +3,10 @@ set -e
 GOEXEC=${GOEXEC:-"go"}
 
 # clean
+if [ -z "$output_dir" ]; then
+  echo "output_dir is empty"
+  exit 1
+fi
 rm -rf $output_dir/bin/ && mkdir -p $output_dir/bin/
 rm -rf $output_dir/log/ && mkdir -p $output_dir/log/
